fix(zod): fix zod leak to client bundle
