+++ src/constants/organization.ts
@@ -1,0 +1,4 @@
+export const OrganizationStatus = {
+  Active: "active",
+  Inactive: "inactive",
+};
