+++ src/types/list.ts
@@ -1,4 +1,4 @@
-import { EventScale } from "./event";
+import { EventScale } from "@/constants/event";
 
 export enum DurationType {
   Passed = "passed", //already done.
