+++ src/types/region.ts
@@ -1,18 +1,5 @@
+import { RegionType } from "@/constants/region";
 import * as z from "zod/v4";
-
-export enum RegionType {
-  COUNTRY = "country",
-  STATE = "state",
-  CITY = "city",
-  DISTRICT = "district",
-}
-
-export const RegionTypeLabel = {
-  [RegionType.COUNTRY]: "国家",
-  [RegionType.STATE]: "省份",
-  [RegionType.CITY]: "城市",
-  [RegionType.DISTRICT]: "区县",
-};
 
 export const RegionSchema = z.object({
   id: z.uuid(),
